burn-import: Refactor test
